fix: 프로젝트 지원중복제거
