+++ src/main/java/com/codevumc/codev_backend/service/co_project/CoProjectService.java
@@ -12,5 +12,4 @@ public interface CoProjectService {
     void updateMainImg(String co_mainImg, long co_projectId);
     CoDevResponse getCoProjects(String co_email, String co_locationTag, String co_partTag, String co_keyword, String co_sortingTag, String co_processTag, int limit, int offset, int page);
     CoDevResponse deleteCoProject(String co_email, long co_projectId);
-    CoDevResponse insertCoRecruitOfProject(CoRecruitOfProject coRecruitOfProject);
 }
