+++ src/main/java/com/codevumc/codev_backend/service/co_projectrecruit/CoProjectRecruitService.java
@@ -1,7 +1,9 @@
 package com.codevumc.codev_backend.service.co_projectrecruit;
 
+import com.codevumc.codev_backend.domain.CoRecruitOfProject;
 import com.codevumc.codev_backend.errorhandler.CoDevResponse;
 
 public interface CoProjectRecruitService {
+    CoDevResponse insertCoRecruitOfProject(CoRecruitOfProject coRecruitOfProject);
     CoDevResponse cancelCoRecruitOfProject(String co_email, long co_projectId);
 }
